requirements.txt
Flask
numpy
Pillow
requests
torch
torchvision